requirements.txt
-i https://rpm:5U9GQ18muJgJ@pypi.jangl.com/pypi/
avro
boto3>=1.4.4
bumpversion>=0.5.3
cachetools>=2.0.0
confluent-kafka[avro]<1.8
Django>=2,<3
django-opentracing==1.1.0
django-storages==1.6.3
djangorestframework>=3.9,<3.10
fastavro>=0.14.7
gevent
jaeger-client==4.4.0
opentracing==2.4.0
opentracing_instrumentation==3.3.1
prettyconf==1.2.0
pykafka>=2.7.0,<2.8
pytz>=2015.4
requests>=2.7.0
requests_oauthlib>=0.8.0,<0.9
rpm-confluent-schemaregistry==0.1.1
sentry-sdk
simpleflake>=0.1.5
urllib3
uwsgi
wheel>=0.24.0
